refactor(opportunity-pages): consolidate opportunity detail components for improved structure
- Replaced individual opportunity detail pages with a unified OpportunityDetailContainer for both volunteer and organization roles.
- Streamlined the import paths for shared components, enhancing code organization.
- Removed redundant state management and queries, simplifying the component logic.
- Improved overall readability and maintainability of the opportunity detail views.
