+++ components/layout/shared/MobileTabsSlider.tsx
