+++ components/layout/volunteer/home-page/OpportunityDetail.tsx
@@ -1,9 +1,8 @@
 "use client";
 
-import { PostSidebar } from "../../../shared/PostSidebar";
+import { PostSidebar } from "../../shared/PostSidebar";
 import { PostContent } from "./PostContent";
- 
- 
+
 type Opportunity = {
   _id: string;
   title: string;
@@ -38,35 +37,34 @@ type Opportunity = {
   };
   banner_img?: string;
   external_event_link?: string;
+  email_contact?: string;
+  phone_contact?: string;
   createdAt: Date;
 };
 
 interface OpportunityDetailProps {
   opportunity: Opportunity;
-  userRole?: "volunteer" | "organization";
-  isMentor?: boolean;
+  userRole?: "volunteer" | "organisation";
 }
 
-export function OpportunityDetail({
-  opportunity,
-   isMentor = false,
-}: OpportunityDetailProps) {
-  
-
-   
-
+export function OpportunityDetail({ opportunity }: OpportunityDetailProps) {
   return (
     <div className="space-y-6 md:space-y-8">
       <div className="flex flex-col lg:flex-row gap-6 lg:gap-8">
         {/* Main Content */}
         <div className="flex-1 min-w-0">
-          <PostContent opportunity={opportunity} isMentor={isMentor} />
+          <PostContent opportunity={opportunity} />
         </div>
 
         {/* Sidebar */}
         <div className="hidden lg:block w-[1px] bg-gray-200"></div>
         <div className="lg:w-[350px] lg:flex-shrink-0">
-          <PostSidebar organization_profile={opportunity.organization_profile as unknown as import("@/server/db/interfaces/organization-profile").IOrgnizationPofile} userRole="volunteer" />
+          <PostSidebar
+            organization_profile={
+              opportunity.organization_profile as unknown as import("@/server/db/interfaces/organization-profile").IOrgnizationPofile
+            }
+            userRole="volunteer"
+          />
         </div>
       </div>
     </div>
